refactor: move `colors-block` to `color-block`
