+++ packages/color-block/manifest.json
