+++ packages/color-block/package.json
@@ -5,8 +5,8 @@
     "version": "1.0.0",
     "main": "dist/index.js",
     "scripts": {
-        "serve": "frontify-cli block serve",
-        "deploy": "frontify-cli block deploy",
+        "serve": "frontify-cli block serve --entryPath src/index.ts",
+        "deploy": "frontify-cli block deploy --entryPath src/index.ts",
         "lint": "eslint src",
         "lint:fix": "eslint --fix src",
         "prettier": "prettier --check src",
