+++ packages/color-block/src/components/drops/DropsItem.spec.ct.tsx
