+++ packages/color-block/src/components/drops/DropsItem.tsx
