+++ packages/color-block/src/global.d.ts
