+++ packages/color-block/tailwind.config.js
